egl: Add tracking for color component locations within a pixel
EGL_KHR_lock_surface needs this to for the EGL_BITMAP_PIXEL_*_OFFSET_KHR
queries.  After describing the behavior of EGL_BITMAP_POINTER_KHR and
EGL_BITMAP_PITCH_KHR, the extension spec says:

    "Other properties of the mapped color buffer of a surface (in the
    list above) are invariant, and need be queried only once following
    surface creation."

To me, this implies these should be properties of the EGLconfig instead
of properties of the surface.

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>