egl: Add eglQuerySurface support for EGL_BITMAP_PIXEL_*_OFFSET_KHR
If EGL_KHR_lock_surface is not supported, generate errors.  If the
extension is supported, return whatever value is stored in the EGLConfig
for the surface.

The extension spec says:

    "If <surface> was created with an EGLConfig whose EGL_SURFACE_TYPE
    attribute does not contain EGL_LOCK_SURFACE_BIT_KHR, queries of
    EGL_BITMAP_ORIGIN_KHR, EGL_BITMAP_PIXEL_<x>_OFFSET_KHR and
    EGL_BITMAP_PIXEL_SIZE_KHR return undefined values."

So we won't check that bit in the EGLConfig, and we'll just return
whatever random data (possibly zeros) is stored in the config.

Note that this implementation never supports luminance surfaces, so, per
the spec, we always return zero in that case:

    "If a color component does not exist in the mapped buffer, then the
    bit offset of that component is zero."

Signed-off-by: Ian Romanick <ian.d.romanick@intel.com>