Allow EGL software rendering by building the EGL Gallium driver
We now build the EGL Gallium driver, which allows software
rendering (for usages like EFL applications with wayland_egl
engine on Weston e.g.).

We patch the source to allow building with latest Wayland.
We change the discovery order to make sure the hardware driver
(egl_dri2) will still be used automatically, and the software
driver as a fallback only.

Change-Id: I98d7caebbfbacc845e23b384765fc7a1adfc0a07
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>